requirements.txt
requests==2.31.0
flask==2.3.3
pyyaml==6.0.1
gitpython==3.1.40
python-dateutil==2.8.2 